Title bar strip: date pill restored, unblocked pill opens the day-window menu
From testing feedback on the macOS title bar:

- The date pill leads again (it was dropped as redundant-with-today,
  but it anchors the bar's numbers just like the desktop strip's).
- The unblocked pill is now interactive: the ONE WebkitAppRegion
  no-drag carve-out in the bar, with a three-dot affordance, opening
  the day-window popover for TODAY. The rest of the bar stays a drag
  region.
- The popover is the timeline strip's, extracted rather than copied:
  new src/components/DayWindowMenu.jsx owns the backdrop, the
  Starts/Ends pickers (portaled ClockTimePicker, same reasons as
  before), and the sticky-forward semantics; positioning belongs to
  the caller. SummaryStrip anchors it upward over the timeline
  (absolute bottom-full, unchanged behavior); the title bar portals it
  to body and opens DOWNWARD under the bar (its div is
  overflow-hidden and screen-top), clamped on-screen at narrow widths.

Groundwork for optionally hiding the timeline strip when the title bar
carries it — editing the day window no longer requires the timeline
strip to be visible.
