+++ src/components/DayWindowMenu.jsx
@@ -1,0 +1,95 @@
+import { useState } from 'react';
+import { createPortal } from 'react-dom';
+import { useDayPlannerCtx } from '../context/DayPlannerContext.jsx';
+import { useFeaturesCtx } from '../context/FeaturesContext.jsx';
+import ClockTimePicker from './ClockTimePicker.jsx';
+
+/**
+ * The day-window (Starts/Ends) popover — ONE implementation shared by
+ * SummaryStrip (opens upward over the timeline, absolutely positioned in its
+ * container) and TitlebarSummaryStrip (portaled to body, opening downward
+ * under the macOS title bar). Positioning belongs to the CALLER via
+ * anchorClass/anchorStyle on the popover box; the full-screen backdrop and the
+ * clock-picker portal are handled here.
+ *
+ * Menu edits act on the RESOLVED window (day entry or inherited default) —
+ * what the user sees in the inputs is what gets stored. Setting either bound
+ * is sticky-forward (also becomes the default); Clear affects this day only.
+ *
+ * The clock picker is PORTALED to document.body (established pattern:
+ * ProjectCard, SchedTaskCard): SummaryStrip's container is
+ * pointer-events-none (inherited, which made an in-place picker
+ * click-transparent) and its sticky z-30 stacking context would cap the
+ * picker's z-[90] under root-level chrome like the phone FAB. Unset bounds
+ * seed the clock at 08:00 / 22:00 so it has a hand to start from.
+ */
+export default function DayWindowMenu({ dateStr, onClose, anchorClass = '', anchorStyle }) {
+  const {
+    darkMode, textPrimary, textSecondary, borderClass, cardBg,
+    use24HourClock, isTablet, formatTime,
+  } = useDayPlannerCtx();
+  const { getDayWindow, setDayWindow, clearDayWindow } = useFeaturesCtx();
+
+  // Which bound the clock picker is editing: 'start' | 'stop' | null.
+  const [pickerField, setPickerField] = useState(null);
+
+  const dayWindow = getDayWindow?.(dateStr) ?? null;
+
+  const updateWindow = (field, value) => {
+    setDayWindow(dateStr, {
+      start: field === 'start' ? (value || null) : dayWindow?.start ?? null,
+      stop: field === 'stop' ? (value || null) : dayWindow?.stop ?? null,
+    });
+  };
+
+  // Trigger buttons open the same clock picker used everywhere else in the app
+  // (FrameEditor, reminders, routines) instead of a bare native time input.
+  const timeBtnCls = `px-2.5 py-1 rounded-lg border text-xs font-medium ${darkMode ? 'bg-gray-800 border-gray-600 text-gray-200 hover:bg-gray-700' : 'bg-white border-stone-300 text-stone-800 hover:bg-stone-50'}`;
+
+  return (
+    <>
+      {/* Backdrop closes on any outside tap/click. */}
+      <div className="fixed inset-0 z-40 pointer-events-auto" onClick={onClose} />
+      <div
+        className={`${anchorClass} z-50 pointer-events-auto w-72 rounded-lg border shadow-xl p-3 space-y-2.5 text-xs ${cardBg} ${borderClass}`}
+        style={anchorStyle}
+      >
+        <div className={`font-semibold ${textPrimary}`}>Day window</div>
+        <p className={textSecondary}>
+          Applies to this and future days. Unblocked time is measured between
+          the <span className={`font-semibold ${textPrimary}`}>Starts</span> and{' '}
+          <span className={`font-semibold ${textPrimary}`}>Ends</span> window.
+        </p>
+        <div className="flex items-center justify-between gap-3">
+          <span className="flex items-center gap-1.5">
+            <span className={textSecondary}>Starts</span>
+            <button onClick={() => setPickerField('start')} className={timeBtnCls}>
+              {dayWindow?.start ? formatTime(dayWindow.start) : 'Set…'}
+            </button>
+          </span>
+          <span className="flex items-center gap-1.5">
+            <span className={textSecondary}>Ends</span>
+            <button onClick={() => setPickerField('stop')} className={timeBtnCls}>
+              {dayWindow?.stop ? formatTime(dayWindow.stop) : 'Set…'}
+            </button>
+          </span>
+        </div>
+        <button
+          onClick={() => { clearDayWindow(dateStr); onClose(); }}
+          className={`w-full py-1 rounded border ${borderClass} ${textSecondary} hover:opacity-80`}
+        >
+          Clear for this day
+        </button>
+      </div>
+      {pickerField && createPortal(
+        <ClockTimePicker
+          value={(pickerField === 'start' ? dayWindow?.start : dayWindow?.stop) ?? (pickerField === 'start' ? '08:00' : '22:00')}
+          onChange={(t) => { updateWindow(pickerField, t); setPickerField(null); }}
+          onClose={() => setPickerField(null)}
+          darkMode={darkMode} isTablet={isTablet ?? false} use24HourClock={use24HourClock}
+        />,
+        document.body,
+      )}
+    </>
+  );
+}
