+++ src/components/SummaryStrip.jsx
@@ -1,9 +1,8 @@
 import { useMemo, useState } from 'react';
-import { createPortal } from 'react-dom';
 import { ChevronDown, ChevronUp, Leaf, MoreHorizontal, Zap } from 'lucide-react';
 import { useDayPlannerCtx } from '../context/DayPlannerContext.jsx';
 import { useFeaturesCtx } from '../context/FeaturesContext.jsx';
-import ClockTimePicker from './ClockTimePicker.jsx';
+import DayWindowMenu from './DayWindowMenu.jsx';
 import { dateToString, formatShortDate } from '../utils/taskUtils.js';
 import { computeDaySummary, formatMinutes } from '../utils/daySummary.js';
 
@@ -61,11 +60,10 @@ export const summaryPillClass = (darkMode) =>
 export default function SummaryStrip({ phone = false, staticPlacement = false }) {
   const {
     selectedDate, getTasksForDate, listEndOfDayTime,
-    darkMode, textPrimary, textSecondary, borderClass, cardBg,
-    use24HourClock, isTablet, formatTime,
+    darkMode, textPrimary, textSecondary,
   } = useDayPlannerCtx();
   const {
-    getDayWindow, setDayWindow, clearDayWindow,
+    getDayWindow,
     dayWindowMenuOpen: menuOpen, setDayWindowMenuOpen: setMenuOpen,
   } = useFeaturesCtx();
 
@@ -79,9 +77,6 @@ export default function SummaryStrip({ phone = false, staticPlacement = false })
     });
   };
 
-  // Which bound the clock picker is editing: 'start' | 'stop' | null.
-  const [pickerField, setPickerField] = useState(null);
-
   const dateStr = dateToString(selectedDate);
   const dayWindow = getDayWindow?.(dateStr) ?? null;
 
@@ -94,20 +89,6 @@ export default function SummaryStrip({ phone = false, staticPlacement = false })
   );
 
   const pill = summaryPillClass(darkMode);
-
-  // Menu edits act on the RESOLVED window (day entry or inherited default) —
-  // what the user sees in the inputs is what gets stored. Setting either bound
-  // is sticky-forward (also becomes the default); Clear affects this day only.
-  const updateWindow = (field, value) => {
-    setDayWindow(dateStr, {
-      start: field === 'start' ? (value || null) : dayWindow?.start ?? null,
-      stop: field === 'stop' ? (value || null) : dayWindow?.stop ?? null,
-    });
-  };
-
-  // Trigger buttons open the same clock picker used everywhere else in the app
-  // (FrameEditor, reminders, routines) instead of a bare native time input.
-  const timeBtnCls = `px-2.5 py-1 rounded-lg border text-xs font-medium ${darkMode ? 'bg-gray-800 border-gray-600 text-gray-200 hover:bg-gray-700' : 'bg-white border-stone-300 text-stone-800 hover:bg-stone-50'}`;
 
   const unblockedLabel = (
     <span className="flex items-baseline gap-1">
@@ -143,55 +124,11 @@ export default function SummaryStrip({ phone = false, staticPlacement = false })
   return (
     <div className={container}>
       {menuOpen && (
-        <>
-          {/* Backdrop closes on any outside tap/click. */}
-          <div className="fixed inset-0 z-40 pointer-events-auto" onClick={() => setMenuOpen(false)} />
-          <div className={`absolute bottom-full mb-1 left-2 z-50 pointer-events-auto w-72 rounded-lg border shadow-xl p-3 space-y-2.5 text-xs ${cardBg} ${borderClass}`}>
-            <div className={`font-semibold ${textPrimary}`}>Day window</div>
-            <p className={textSecondary}>
-              Applies to this and future days. Unblocked time is measured between
-              the <span className={`font-semibold ${textPrimary}`}>Starts</span> and{' '}
-              <span className={`font-semibold ${textPrimary}`}>Ends</span> window.
-            </p>
-            <div className="flex items-center justify-between gap-3">
-              <span className="flex items-center gap-1.5">
-                <span className={textSecondary}>Starts</span>
-                <button onClick={() => setPickerField('start')} className={timeBtnCls}>
-                  {dayWindow?.start ? formatTime(dayWindow.start) : 'Set…'}
-                </button>
-              </span>
-              <span className="flex items-center gap-1.5">
-                <span className={textSecondary}>Ends</span>
-                <button onClick={() => setPickerField('stop')} className={timeBtnCls}>
-                  {dayWindow?.stop ? formatTime(dayWindow.stop) : 'Set…'}
-                </button>
-              </span>
-            </div>
-            <button
-              onClick={() => { clearDayWindow(dateStr); setMenuOpen(false); }}
-              className={`w-full py-1 rounded border ${borderClass} ${textSecondary} hover:opacity-80`}
-            >
-              Clear for this day
-            </button>
-          </div>
-          {/* Same clock picker as FrameEditor/reminders/routines. PORTALED to
-              document.body (established pattern: ProjectCard, SchedTaskCard),
-              for two reasons this container creates: pointer-events-none is
-              inherited, which made an in-place picker click-transparent (taps
-              fell through to the menu backdrop and closed everything), and the
-              sticky z-30 stacking context would cap the picker's z-[90] below
-              root-level chrome like the phone FAB. Unset bounds seed the clock
-              at 08:00 / 22:00 so it has a hand to start from. */}
-          {pickerField && createPortal(
-            <ClockTimePicker
-              value={(pickerField === 'start' ? dayWindow?.start : dayWindow?.stop) ?? (pickerField === 'start' ? '08:00' : '22:00')}
-              onChange={(t) => { updateWindow(pickerField, t); setPickerField(null); }}
-              onClose={() => setPickerField(null)}
-              darkMode={darkMode} isTablet={isTablet ?? false} use24HourClock={use24HourClock}
-            />,
-            document.body,
-          )}
-        </>
+        <DayWindowMenu
+          dateStr={dateStr}
+          onClose={() => setMenuOpen(false)}
+          anchorClass="absolute bottom-full mb-1 left-2"
+        />
       )}
       {isEmptyHint ? (
         <button
